refactor: 重构代码 feat：环境配置支持持久化到本地，不用每次都要频繁切换环境了

+++ src/main/kotlin/com/github/qianmi/domain/project/tools/ShellElement.kt
@@ -25,7 +25,7 @@ class ShellElement {
     /**
      * 端口
      */
-    private fun getPort(): Int = ShellConfig.getInstance().port
+    private fun getPort(): Int = ShellConfig.getInstance().port.toInt()
 
     /**
      * 工作路径
